requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.3.10
backoff==2.2.1
bcrypt==4.2.0
bitsandbytes==0.43.2
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
dill==0.4.0
distro==1.9.0
dnspython==2.6.1
duckduckgo_search==6.2.5
einops==0.8.0
email_validator==2.2.0
fake-useragent==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.15.4
flash-attn==2.6.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
funcparserlib==1.0.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.65.1
h11==0.14.0
hnfc-data-mgmt @ git+http://gitlab/my-packages/hnfc-data-mgmt.git@8db9c80a177cf80b7d1397610e4b0a608fa2869f
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
isort==6.0.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
kubernetes==30.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.84.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.5.0
posthog==3.5.0
primp==0.5.5
protobuf==4.25.4
psutil==6.0.0
psycopg2==2.9.9
py-readability-metrics==1.4.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pylint==3.3.7
pynvml==11.5.3
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-whois==0.9.4
pytz==2024.2
PyYAML==6.0.1
QwantAI==1.0.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruff==0.6.7
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@46350f5eae87ac1d168ddfdc57a0b39b64b9a029
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.6.0
websockets==12.0
wrapt==1.16.0
zipp==3.19.2
